add line_path to solution and make compile
