+++ src/gui/gui_generate_qr.py
@@ -6,7 +6,8 @@ from tkinter.messagebox import showerror
 from qrcodegen import QrCode
 
 from src.resources import app_image_path
-from src.platform.vectorize_qr import VectorizeQr
+from src.platform.vectorize_helper import QrValueTable
+from src.platform.line_path import LinePath
 
 
 class GuiGenerateQr:
@@ -14,18 +15,17 @@ class GuiGenerateQr:
         self._main = main
         self._options = options
 
-        self._qr = None
-        self._spiral_path = None
+        self._path = None
         self._stop_draw = False
         self._turtle = None
 
         self.tc_frame = self._init_frame_text_convert()
         self.dwg_frame = self._init_frame_turtle()
 
-    def get_qr_spiral_paths(self):
+    def get_qr_path(self):
         """Getter method.
         :returns the spiral path if created, otherwise None."""
-        return self._spiral_path
+        return self._path
 
     def _init_frame_text_convert(self):
         """Initializes the Text-to-QR-code generator section of the GUI"""
@@ -74,32 +74,34 @@ class GuiGenerateQr:
     def _create_qr_from_input(self, text_to_qr):
         """This method requests a QR code to be generated by the library.
         It then vectorizes it and generates spiral paths."""
-        self._qr = QrCode.encode_text(text_to_qr, QrCode.Ecc.MEDIUM)
-        qr_vect = VectorizeQr(self._qr, 0)
-        self._spiral_path = qr_vect.generate_spiral_path()
+        qr = QrCode.encode_text(text_to_qr, QrCode.Ecc.MEDIUM)
+        qr_vect = QrValueTable()
+        qr_vect.set_qr(qr)
+        self._path = LinePath(qr_vect)
 
     def _draw_qr_turtle(self):
         """Method that draws a QR code path based on the QrPathSegment data class with Turtle."""
         self._main.update_status('\u270d Drawing')
 
-        path_count = len(self._spiral_path)
+        path_count = len(self._path.get_vectors())
         self.progress.config(maximum=path_count)
         self._stop_draw = False
         for i in range(path_count):
             self.progress.step()
             if self._stop_draw:
                 break
-            for vect in self._spiral_path[i].get_z_vector():
+            for vect in self._path.get_vectors():
                 if self._stop_draw:
                     break
-                length = vect.get_length()
-                if vect.get_state():
-                    self._turtle.down()
-                    self._turtle.forward(self.pen_size * length)
-                    self._turtle.up()
+                if vect.x_length == 0:
+                    length = vect.y_length
                 else:
-                    self._turtle.forward(self.pen_size * length)
-            self._turtle.right(90)
+                    length = vect.x_length
+                self._turtle.setpos(self.pen_size * vect.position.x, self.pen_size * vect.position.y)
+                self._turtle.setheading(vect.x_length * 90)
+                self._turtle.down()
+                self._turtle.forward(self.pen_size * length)
+                self._turtle.up()
 
         self._turtle.hideturtle()
         self._stop_draw = False
@@ -150,7 +152,7 @@ class GuiGenerateQr:
 
         self._create_qr_from_input(text)
         self._main.set_project_name(text)
-        self._prepare_screen_for_drawing(self._qr.get_size())
+        self._prepare_screen_for_drawing(self._path.get_size())
         self._draw_qr_turtle()
 
     def _stop_draw_button_clicked(self):
