+++ src/gui/gui_xy0_configure.py
@@ -3,7 +3,7 @@ from tkinter import ttk
 
 from src.helpers.gui_helpers import validate_number
 from src.resources import app_icon_path
-from src.platform.vectorize_qr import Point
+from src.platform.vectorize_helper import Point
 
 
 class Offset:
