+++ test/test_vectorize_qr.py
@@ -2,7 +2,7 @@ import unittest
 from unittest.mock import MagicMock
 
 from qrcodegen import QrCode
-from src.platform.vectorize_qr import Direction, Point, Line, VectorizeQr, create_spiral
+from src.platform.vectorize_helper import Direction, Point, Line, VectorizeQr, create_spiral
 
 
 class TestLine(unittest.TestCase):
